fix(clamshell): directly execute actions
This commit refactors the regolith-sway-clamshell script to directly
execute the action that is to be performed, instead of setting the
binding. This fixes issues with the binding not staying up to date
with the power mode and display configuration of the system.

A new argument, ACTION_TYPE, is introduced, which can be either
LID_OPEN or LID_CLOSE. The script now handles these two events
differently, enabling or disabling the inbuilt display as appropriate.

The sway config has been updated to call the script with the
appropriate ACTION_TYPE when the lid is opened or closed.