chore: remove the audio idle inhibit package
Currently, we start sway-audio-idle-inhibit using a sway config file.
However, this doesn't recover the service from failures and doesn't
allow for checking the status / logs. To improve on this,
https://github.com/regolith-linux/SwayAudioIdleInhibit/pull/6 introduces
a systemd service that handles the execution and management for the
program binary.

This change supplements the systemd PR by removing the config file
and it's corresponding package that handling execution of
sway-audio-idle-inhibit.